requirements.txt
numpy
pytorch
matplotlib
pandas
